requirements.txt
PyYAML
matplotlib
numpy
scipy
PyQt6
